+++ packages/svelte/tests/compiler-errors/samples/export-not-defined-module/_config.js
@@ -1,0 +1,9 @@
+import { test } from '../../test';
+
+export default test({
+	error: {
+		code: 'export_undefined',
+		message: '`blah` is not defined',
+		position: [26, 30]
+	}
+});