+++ packages/svelte/tests/css/samples/snippets-elements/input.svelte
@@ -1,0 +1,13 @@
+{#snippet foo()}
+	<x>
+		<y></y>
+	</x>
+{/snippet}
+
+{@render foo()}
+
+<style>
+	x y {
+		color: green;
+	}
+</style>